backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
langchain
langchain_perplexity